backend/requirements.txt
# Backend dependencies

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4

# DashScope (Alibaba Cloud Model Service)
dashscope>=1.14.0

# LangGraph / LangChain
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0

# Rendering and export
playwright>=1.40.0

# Async file operations
aiofiles>=23.2.0

# WebSocket
websockets>=12.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
